// src/models/products.ts
export default interface IProduct {
  id: string;
  name: string;
  category: string;
  brand: string;
  release: string;
  color?: string;
  model: string;
  price: string;
  storage?: string;
  imgs: string[];
  rating: number;
  stock: number;
  displaySize?: string;
  display?: string;
  resolution?: string;
  ppi?: string;
  refreshRate?: string;
  communication?: string;
  features?: string;
  batteryType?: string;
  sdCardSupport?: string;
  internet?: string;
  numOfSIMs?: string;
  typeOfSIM?: string;
  chipset?: string;
  cameras?: string;
  maxVideoRecording?: string;
  bluetooth?: string;
  nfc?: string;
  os?: string;
  osVersion?: string;
  protection?: string;
  material?: string;
  size?: string;
  weight: string;
  earpieceDesign?: string;
  construction?: string;
  connection?: string;
  activeNoiseCancellation?: string;
  numOfMics?: string;
  audioCodec?: string;
  caseWeight?: string;
  coating?: string;
  chipsetCores?: string;
  ram?: string;
  storageType?: string;
  graphicsCardType?: string;
  wifiStandart?: string;
  ports?: string;
  webCamera?: string;
  training?: string;
  bodyShape?: string;
  control?: string;
  sensors?: string;
  bandColor?: string;
  caseSize?: string;
}